+++ react/ProductAssemblyOptions.tsx
@@ -4,7 +4,7 @@ import useAssemblyOptions from './modules/useAssemblyOptions'
 import ProductAssemblyOptionsGroup from './components/ProductAssemblyOptions/ProductAssemblyOptionsGroup'
 import { ProductAssemblyGroupContextProvider } from './components/ProductAssemblyContext/Group'
 
-const ProductAssemblyOptions: FC = ({ children }) => {
+const ProductAssemblyOptions: FC<AssemblyOptionsGroup> = ({ children, initiallyOpened = false }) => {
   const assemblyOptions = useAssemblyOptions()
 
   if (!assemblyOptions) {
@@ -18,7 +18,7 @@ const ProductAssemblyOptions: FC = ({ children }) => {
           key={assemblyOptionId}
           assemblyOption={assemblyOptions[assemblyOptionId]}
         >
-          <ProductAssemblyOptionsGroup>{children}</ProductAssemblyOptionsGroup>
+          <ProductAssemblyOptionsGroup initiallyOpened={initiallyOpened}>{children}</ProductAssemblyOptionsGroup>
         </ProductAssemblyGroupContextProvider>
       ))}
     </Fragment>